requirements.txt
flask

Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask_Cors
Flask-qrcode
Flask-Minify
Flask-Moment

werkzeug
pyjwt
pyotp

click
hashlib

requests
IMDbPY
beautifulsoup4
